rename of files and associated changes throughout components
