+++ npm-debug.log
@@ -1,0 +1,48 @@
+0 info it worked if it ends with ok
+1 verbose cli [ 'C:\\node.exe',
+1 verbose cli   'C:\\node_modules\\npm\\bin\\npm-cli.js',
+1 verbose cli   'run',
+1 verbose cli   'build' ]
+2 info using npm@3.10.10
+3 info using node@v6.11.2
+4 verbose run-script [ 'prebuild', 'build', 'postbuild' ]
+5 info lifecycle joystick-informer-client@0.1.0~prebuild: joystick-informer-client@0.1.0
+6 silly lifecycle joystick-informer-client@0.1.0~prebuild: no script for prebuild, continuing
+7 info lifecycle joystick-informer-client@0.1.0~build: joystick-informer-client@0.1.0
+8 verbose lifecycle joystick-informer-client@0.1.0~build: unsafe-perm in lifecycle true
+9 verbose lifecycle joystick-informer-client@0.1.0~build: PATH: C:\node_modules\npm\bin\node-gyp-bin;C:\Users\Zak\desktop\thinkful\joystick-informer\joystick-informer\node_modules\.bin;C:\Users\Zak\bin;C:\Program Files\Git\mingw64\bin;C:\Program Files\Git\usr\local\bin;C:\Program Files\Git\usr\bin;C:\Program Files\Git\usr\bin;C:\Program Files\Git\mingw64\bin;C:\Program Files\Git\usr\bin;C:\Users\Zak\bin;C:\ProgramData\Oracle\Java\javapath;C:\Program Files\Common Files\Microsoft Shared\Windows Live;C:\Program Files (x86)\Common Files\Microsoft Shared\Windows Live;C:\Program Files (x86)\Intel\iCLS Client;C:\Program Files\Intel\iCLS Client;C:\Windows\system32;C:\Windows;C:\Windows\System32\Wbem;C:\Windows\System32\WindowsPowerShell\v1.0;C:\Program Files\Intel\Intel(R) Management Engine Components\DAL;C:\Program Files\Intel\Intel(R) Management Engine Components\IPT;C:\Program Files (x86)\Intel\Intel(R) Management Engine Components\DAL;C:\Program Files (x86)\Intel\Intel(R) Management Engine Components\IPT;C:\Program Files (x86)\ATI Technologies\ATI.ACE\Core-Static;C:\Program Files (x86)\QuickTime\QTSystem;C:\Program Files (x86)\Windows Live\Shared;C:\Program Files (x86)\AMD\ATI.ACE\Core-Static;C:\Python27;C:\Program Files (x86)\NVIDIA Corporation\PhysX\Common;C:\Program Files\Git\cmd;C:\Program Files (x86)\Windows Kits\8.1\Windows Performance Toolkit;C:\WINDOWS\system32;C:\WINDOWS;C:\WINDOWS\System32\Wbem;C:\WINDOWS\System32\WindowsPowerShell\v1.0;C:\Program Files (x86)\Git\cmd;C:\;C:\Ruby24-x64\bin;C:\Users\Zak\AppData\Local\Microsoft\WindowsApps;C:\Program Files\Sublime Text 3;C:\Program Files\Heroku;C:\Program Files\Heroku\bin;C:\Program Files\MongoDB\Server\3.2\bin;C:\Users\Zak\AppData\Roaming\npm;C:\Program Files\Git\usr\bin\vendor_perl;C:\Program Files\Git\usr\bin\core_perl
+10 verbose lifecycle joystick-informer-client@0.1.0~build: CWD: C:\Users\Zak\desktop\thinkful\joystick-informer\joystick-informer
+11 silly lifecycle joystick-informer-client@0.1.0~build: Args: [ '/d /s /c', 'react-scripts build' ]
+12 silly lifecycle joystick-informer-client@0.1.0~build: Returned: code: 1  signal: null
+13 info lifecycle joystick-informer-client@0.1.0~build: Failed to exec build script
+14 verbose stack Error: joystick-informer-client@0.1.0 build: `react-scripts build`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (C:\node_modules\npm\lib\utils\lifecycle.js:255:16)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at EventEmitter.emit (events.js:191:7)
+14 verbose stack     at ChildProcess.<anonymous> (C:\node_modules\npm\lib\utils\spawn.js:40:14)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at ChildProcess.emit (events.js:191:7)
+14 verbose stack     at maybeClose (internal/child_process.js:891:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
+15 verbose pkgid joystick-informer-client@0.1.0
+16 verbose cwd C:\Users\Zak\desktop\thinkful\joystick-informer\joystick-informer
+17 error Windows_NT 10.0.15063
+18 error argv "C:\\node.exe" "C:\\node_modules\\npm\\bin\\npm-cli.js" "run" "build"
+19 error node v6.11.2
+20 error npm  v3.10.10
+21 error code ELIFECYCLE
+22 error joystick-informer-client@0.1.0 build: `react-scripts build`
+22 error Exit status 1
+23 error Failed at the joystick-informer-client@0.1.0 build script 'react-scripts build'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the joystick-informer-client package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     react-scripts build
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs joystick-informer-client
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls joystick-informer-client
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
