+++ src/components/relatedGames.js
@@ -55,9 +55,9 @@ export class RelatedGames extends React.Component {
 
 const mapStateToProps = state => {
 	return {
-		loggedIn: state.joystick.currentUser !== null,
-		currentUser: state.joystick.currentUser,
-		currentWatchlist: state.joystick.currentWatchlist
+		loggedIn: state.auth.currentUser !== null,
+		currentUser: state.auth.currentUser,
+		currentWatchlist: state.auth.currentWatchlist
 	}
 };
