+++ src/components/watchlist.js
@@ -43,9 +43,9 @@ export class Watchlist extends React.Component {
 
 const mapStateToProps = state => {
 	return {
-		loggedIn: state.joystick.currentUser !== null,
-		currentUser: state.joystick.currentUser,
-		currentWatchlist: state.joystick.currentWatchlist
+		loggedIn: state.auth.currentUser !== null,
+		currentUser: state.auth.currentUser,
+		currentWatchlist: state.auth.currentWatchlist
 	}
 };
