+++ src/store.js
@@ -3,7 +3,7 @@ import {devToolsEnhancer} from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 
 import {reducer as formReducer} from 'redux-form';
-import {joystickReducer} from './reducers';
+import {joystickReducer} from './reducers/joystick';
 import {authReducer} from './reducers/auth';
 
 const combinedReducer = combineReducers({
